Make 'init' compatible with the new namespace packages.
After some internal discussions we agreed that package names alone are
somewhat restrictive as the sole unique identifier. We agreed it would
be best to support a model similar to Java (gasp!) and use the reverse
DNS syntax as a prefix to the package name. This is the first step.