Re-factor the install command to parse package.json files.
The existing version of the install command used the dependency list it
got from the server to build up the directories and generate the
vendor.pri file. This new version uses the actual package data by
parsing the package file for each dependency. This gives us more
information about the dependencies at install time so we can be more
clever in the future. One example here is if we want package.json to be
able to specify a custom pri file.