Implement a new version of uninstall.
Instead of using the package file as the truth for which dependencies
exist in the vendorDir, we now trust the file system instead. We search
the vendorDir for package files and build up a map to see which ones
exist. We can then remove the requested package and update the vendor.pri
file. If the current directory contains a package file, we will also
update that by removing the specified dependency.