Attempt to fix GCS deploy by tweaking local-dir.
The build failed because the dpl module appeared to have trouble finding
the 'bin' directory. This might be caused by the multiple paths in
the GOPATH environment variable. To try to fix this, we set the local-dir
to point to the first path in GOPATH and hope that it finds the stuff
there.