Add some docs about package naming.
For packages that wrap an existing technology, it can be tempting to
prefix the package name with "qt" or "qml". We should discourage this
behaviour since in the context of qpm, it is superfluous. Everything is
Qt or QML :) Also, mention the namespacing and why we did it like that.