Change the format of Repository.Url to store full clone URLs.
Storing the Repository.Url as a GitHub specific sub URL makes little
sense to users. We are better off to store the full clone URL in the
database and then update the install command to extract the parts it
needs. This will require we update the existing packages which is why we
are doing it sooner rather than later.