Extend the DependencyResponse with a list of server messages.
This adds support for a fairly robust messaging system where the server
can inform or warn the user about various things when installing packages.
The first use of this will most likely be to warn about potential license
incompatibilities, but it could also be used for deprecation warnings in
the future or conflicting versions, etc.

Having the logic for this on the server allows us to minimise the logic
in the client and therefore worry less about outdated versions of the
client.