Send the dependant's license in the dependency request.
Now that we are adding support for more (restrictive) licenses, we want
to avoid the case where people start blindly pulling in dependencies
without understanding the consequences. For example, pulling in a strong
copyleft licensed dependency could have serious implications. To avoid
this, we send the license of the dependant package (eg: the app) to the
server so the server can return warnings before continuing.