Improve "qpm list/search" with better data and formatting.
The existing implementation of this function had gotten many complaints
since it did not print the description which is arguably the most
important piece of info. Also, it completely ignored the width of the
screen so wrapping completely broke the formatting. This new version
is "responsive" to the screen width and prints more data if it is
available.

Fixes #8.