Add support for qbs files
Upon install a vendor.qbs file with a Project Item containing a
Product Item named "vendor" is created.
The "vendor" project can then be added as Depends { name: "vendor" }
in the client project.
The Product Item named "vendor" depends on all the provided
packages.
The qpm.json file should contain a "qbs_filename" property
and package.qbs should be similar to the following:

    import qbs 1.0

    Product {
        name: "net.ovilab.SimVis"
        type: "staticlibrary"

        Depends { name: 'cpp' }
        Depends { name: "Qt.core" }

        files: [
            "myfile.h",
            "myfile.cpp"
        }
    }

Due to technical limitations making it impossible to add a Group Item
inside another Group Item, a Product with type: "staticlibrary" is used.