Always use Unix path seperators when splitting strings from the tarball
The Github tarballs (maybe all?) use Unix path seperators to specify
path names so we should not try to split the path using OS native path
seperators since this will not split anything on Windows. As a result of
this, the 'topDir' variable could contain a nested path which obviously
doesn't contain a qpm.json file and thus resulting in an error.

Fixes #24.